+++ src/components/Table/Table.tsx
@@ -1,4 +1,4 @@
-// Tremor Raw Table [v0.0.1]
+// Tremor Raw Table [v0.0.2]
 
 import React from "react"
 
@@ -10,7 +10,7 @@ const TableRoot = React.forwardRef<
 >(({ className, children, ...props }, forwardedRef) => (
   <div
     ref={forwardedRef}
-    // Activate if table is used in a float enironment
+    // Activate if table is used in a float environment
     // className="flow-root"
   >
     <div