+++ src/components/Calendar/Calendar.tsx
@@ -124,7 +124,8 @@ const Calendar = ({
           "text-gray-900 dark:text-gray-50",
         ),
         day: cx(
-          "size-9 rounded text-sm text-gray-900 dark:text-gray-50",
+          "size-9 rounded text-sm focus:z-10",
+          "text-gray-900 dark:text-gray-50",
           "hover:bg-gray-200 hover:dark:bg-gray-700",
           focusRing,
         ),