+++ src/components/Calendar/changelog.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - Fix: Z-index day
+- Fix: showOutsideDays logic
 
 ## 0.0.3