+++ src/components/Calendar/changelog.md
@@ -4,6 +4,7 @@
 
 ### Changes
 
+- Fix: Z-index day
 - Fix: showOutsideDays logic
 
 ## 0.0.3